[Network Traffic] Using hpp for headers.
