+++ tools/network-traffic/src/DummyServer.cpp
@@ -1,4 +1,4 @@
-#include "DummyServer.h"
+#include "DummyServer.hpp"
 
 #include <WS2tcpip.h>
